pyproject.toml
[tool.poetry]
name = "harkai"
version = "0.1.0"
description = ""
authors = ["Chetan Malhotra <malhotra_chetan@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-texttospeech = "^2.16.5"
pydub = "^0.25.1"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
